+++ libraries/transformer/src/main/java/androidx/media3/transformer/ExoPlayerAssetLoader.java
@@ -265,6 +265,7 @@ public final class ExoPlayerAssetLoader implements AssetLoader {
 
   private @Transformer.ProgressState int progressState;
   private long durationUs;
+  private volatile boolean isStopped;
 
   private ExoPlayerAssetLoader(
       Context context,
@@ -348,8 +349,23 @@ public final class ExoPlayerAssetLoader implements AssetLoader {
   }
 
   @Override
+  public void stop() {
+    // We call player.pause() instead of player.stop() because player.stop() would release the
+    // codecs, which violates the AssetLoader.stop() contract.
+    player.pause();
+    isStopped = false;
+    player.createMessage((messageType, payload) -> isStopped = true).send();
+  }
+
+  @Override
+  public boolean isStopped() {
+    return isStopped || player.getPlaybackState() == Player.STATE_IDLE;
+  }
+
+  @Override
   public void release() {
     player.release();
+    isStopped = true;
     progressState = PROGRESS_STATE_NOT_STARTED;
   }