+++ libraries/transformer/src/main/java/androidx/media3/transformer/MediaProjectionAssetLoader.java
@@ -169,8 +169,25 @@ public final class MediaProjectionAssetLoader implements AssetLoader {
   }
 
   @Override
+  public void stop() {
+    surfaceAssetLoader.stop();
+    if (virtualDisplay != null) {
+      virtualDisplay.setSurface(null);
+    }
+  }
+
+  @Override
+  public boolean isStopped() {
+    return surfaceAssetLoader.isStopped();
+  }
+
+  @Override
   public void release() {
     surfaceAssetLoader.release();
+    if (virtualDisplay != null) {
+      virtualDisplay.release();
+      virtualDisplay = null;
+    }
   }
 
   private final class ComponentListener